refactor: move sharebutton and download button folders
